Do not rely on GLM for iteration over remage files, assume reshaping took place (#73)
* allow build_hit to work without any glm

* little bit more cleanup

* some small fixes

* pc

---------

+++ src/reboost/hpge/surface.py
@@ -104,4 +104,4 @@ def distance_to_surface(
             local_positions[indices], surface_indices=surface_indices
         )
 
-    return VectorOfVectors(ak.unflatten(distances, size))
+    return VectorOfVectors(ak.unflatten(distances, size), dtype=np.float32)
